% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-exposure-groups.R
\name{analyze_exposure_groups}
\alias{analyze_exposure_groups}
\title{Analyze results for the overall exposure groups}
\usage{
analyze_exposure_groups(
  combined_results,
  data,
  outcome,
  exposures,
  mixture_fn,
  refit_mixtures = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{combined_results}{tbd}

\item{data}{tbd}

\item{outcome}{tbd}

\item{exposures}{tbd}

\item{mixture_fn}{tbd}

\item{refit_mixtures}{tbd}

\item{verbose}{tbd}
}
\description{
Not relevant if there is only a single mixture
}
