% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-folds.R
\name{analyze_folds}
\alias{analyze_folds}
\title{This function will be called from tlmixture() on each CV-TMLE split.}
\usage{
analyze_folds(splits, outcome, exposures, family, quantiles_mixtures,
  quantiles_exposures, estimator_outcome, estimator_propensity,
  cluster_exposures, sl_folds = 2L, verbose = FALSE)
}
\description{
This function will be called from tlmixture() on each CV-TMLE split.
}
