% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-exposure-weights.R
\name{create_exposure_weights}
\alias{create_exposure_weights}
\title{Create exposure weights}
\usage{
create_exposure_weights(data, outcome, exposures, quantiles,
  verbose = FALSE)
}
\arguments{
\item{data}{tbd}

\item{outcome}{tbd}

\item{exposures}{tbd}

\item{quantiles}{tbd}

\item{verbose}{tbd}
}
\description{
Create exposure weights
}
