% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-backfit-sl.R
\name{mixture_backfit_sl}
\alias{mixture_backfit_sl}
\title{Create exposure weights via SL with backfitting}
\usage{
mixture_backfit_sl(
  data,
  outcome,
  exposures,
  estimator_mixture = c("SL.mean", "SL.lm"),
  estimator_confounders = estimator_confounders,
  folds_mixture = 5L,
  folds_confounders = 5L,
  exposure_groups = NULL,
  max_iterations = 10L,
  tolerance = 1e-05,
  quantiles = NULL,
  family = "gaussian",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{tbd}

\item{outcome}{outcome column name}

\item{exposures}{tbd}

\item{estimator_mixture}{tbd}

\item{estimator_confounders}{tbd}

\item{exposure_groups}{List of all exposure groups}

\item{max_iterations}{tbd}

\item{tolerance}{tbd}

\item{quantiles}{tbd}

\item{family}{Not used for this mixture estimator.}

\item{verbose}{tbd}

\item{...}{tbd}
}
\description{
Create exposure weights via SL with backfitting
}
