% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-glm.R
\name{mixture_glm}
\alias{mixture_glm}
\title{Create exposure weights via glm}
\usage{
mixture_glm(
  data,
  outcome,
  exposures,
  quantiles,
  family = gaussian(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{tbd}

\item{outcome}{outcome column name}

\item{exposures}{tbd}

\item{quantiles}{tbd}

\item{family}{tbd}

\item{verbose}{tbd}

\item{...}{tbd}
}
\description{
Create exposure weights via glm
}
