% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-lm.R
\name{mixture_lm}
\alias{mixture_lm}
\title{Create exposure weights via lm}
\usage{
mixture_lm(
  data,
  outcome,
  exposures,
  exposure_groups,
  quantiles,
  family = gaussian(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{tbd}

\item{outcome}{outcome column name}

\item{exposures}{tbd}

\item{exposure_groups}{List of all exposure groups}

\item{quantiles}{tbd}

\item{family}{Not used for this mixture estimator.}

\item{verbose}{tbd}

\item{...}{tbd}
}
\description{
Create exposure weights via lm
}
