% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-sl.R
\name{mixture_sl}
\alias{mixture_sl}
\title{Create exposure weights}
\usage{
mixture_sl(
  data,
  outcome,
  exposures,
  quantiles,
  family = "binomial",
  verbose = FALSE,
  sl_library = c("SL.mean", "SL.glm")
)
}
\arguments{
\item{data}{tbd}

\item{outcome}{outcome column name}

\item{exposures}{tbd}

\item{quantiles}{tbd}

\item{family}{tbd}

\item{verbose}{tbd}

\item{sl_library}{tbd}
}
\description{
Create exposure weights
}
