% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome-family.R
\name{outcome_family}
\alias{outcome_family}
\title{Detect outcome family based on the outcome data vector}
\usage{
outcome_family(data, outcome_field)
}
\description{
Detect outcome family based on the outcome data vector
}
