% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-backfit-glm.R
\name{predict.mixture_backfit_glm}
\alias{predict.mixture_backfit_glm}
\title{tbd}
\usage{
\method{predict}{mixture_backfit_glm}(object, data, ...)
}
\arguments{
\item{object}{tbd}

\item{data}{tbd}

\item{...}{tbd}
}
\description{
tbd
}
