% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-backfit-lm.R, R/mixture-backfit-sl.R
\name{predict.mixture_backfit_lm}
\alias{predict.mixture_backfit_lm}
\title{tbd}
\usage{
\method{predict}{mixture_backfit_lm}(object, data, ...)

\method{predict}{mixture_backfit_lm}(object, data, ...)
}
\arguments{
\item{object}{tbd}

\item{data}{tbd}

\item{...}{tbd}
}
\description{
tbd

tbd
}
