% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-backfit-sl.R
\name{predict.mixture_backfit_sl}
\alias{predict.mixture_backfit_sl}
\title{tbd}
\usage{
\method{predict}{mixture_backfit_sl}(object, data, ...)
}
\arguments{
\item{object}{tbd}

\item{data}{tbd}

\item{...}{tbd}
}
\description{
tbd
}
