% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-lm.R
\name{predict.mixture_lm}
\alias{predict.mixture_lm}
\title{tbd}
\usage{
\method{predict}{mixture_lm}(object, data, ...)
}
\arguments{
\item{object}{tbd}

\item{data}{tbd}

\item{...}{tbd}
}
\description{
tbd
}
