% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-pls.R
\name{predict.mixture_pls}
\alias{predict.mixture_pls}
\title{prediction for mixture_pls object}
\usage{
\method{predict}{mixture_pls}(object, data, ...)
}
\arguments{
\item{object}{tbd}

\item{data}{tbd}

\item{...}{tbd}
}
\description{
prediction for mixture_pls object
}
