% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-score-glm.R
\name{predict.mixture_score_glm}
\alias{predict.mixture_score_glm}
\title{tbd}
\usage{
\method{predict}{mixture_score_glm}(object, data, type = "mixture", ...)
}
\arguments{
\item{object}{tbd}

\item{data}{tbd}

\item{type}{"mixture" (default) or "confounders"}

\item{...}{tbd}
}
\description{
tbd
}
