% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-score-sl3.R
\name{predict.mixture_score_sl3}
\alias{predict.mixture_score_sl3}
\title{tbd}
\usage{
\method{predict}{mixture_score_sl3}(object, data, ...)
}
\arguments{
\item{object}{tbd}

\item{data}{tbd}

\item{...}{tbd}

\item{type}{"mixture" (default) or "confounders"}
}
\description{
tbd
}
