% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tlmixture.R
\name{print.tlmixture}
\alias{print.tlmixture}
\title{Prints tlmixture result object}
\usage{
\method{print}{tlmixture}(x, ...)
}
\arguments{
\item{x}{tlmixture result object}

\item{...}{Not used currently.}
}
\description{
Prints tlmixture result object
}
