% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlmixture.R
\name{tlmixture}
\alias{tlmixture}
\title{Targeted learning for exposure mixtures}
\usage{
tlmixture(data, outcome, exposures, quantiles_mixtures = 3L,
  quantiles_exposures = 4L, folds_cvtmle = 2L,
  estimator_outcome = c("SL.mean", "SL.glmnet"),
  estimator_propensity = estimator_outcome, cluster_exposures = FALSE,
  verbose = FALSE)
}
\arguments{
\item{data}{Data frame with outcome, exposure, and adjustment variables.}

\item{outcome}{Name of the outcome variable.}

\item{exposures}{A vector of exposure names, or (not yet supported) a list where each element is
a vector of pre-clustered exposures.}

\item{quantiles_mixtures}{Number of quantiles to use for discretizing mixture
(default 3 - low, medium, high).}

\item{quantiles_exposures}{Number of quantiles to use for discretizing continuous exposures
(default 4).}

\item{folds_cvtmle}{Number of CV-TMLE folds (default 2).}

\item{estimator_outcome}{SuperLearner library for outcome estimation.}

\item{estimator_propensity}{SuperLearner library for propensity estimation.}

\item{cluster_exposures}{Whether to automatically cluster a vector of exposures into
sub-groups (default FALSE, TRUE not yet supported).}

\item{verbose}{If TRUE, display more detailed info during execution.}
}
\description{
This is our main function.
}
