% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-folds.R
\name{analyze_folds}
\alias{analyze_folds}
\title{This function will be called from tlmixture() on each CV-TMLE split.}
\usage{
analyze_folds(
  splits,
  outcome,
  exposures,
  family,
  quantiles_mixtures,
  quantiles_exposures,
  estimator_outcome,
  estimator_propensity,
  mixture_fn,
  cluster_exposures,
  folds_sl = 2L,
  verbose = FALSE
)
}
\arguments{
\item{splits}{tbd}

\item{outcome}{name of the outcome variable}

\item{exposures}{tbd}

\item{family}{tbd}

\item{quantiles_mixtures}{tbd}

\item{quantiles_exposures}{tbd}

\item{estimator_outcome}{tbd}

\item{estimator_propensity}{tbd}

\item{mixture_fn}{tbd}

\item{cluster_exposures}{tbd}

\item{folds_sl}{tbd}

\item{verbose}{tbd}
}
\description{
This function will be called from tlmixture() on each CV-TMLE split.
}
