% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-mixture.R
\name{analyze_mixture}
\alias{analyze_mixture}
\title{Analyze mixture}
\usage{
analyze_mixture(
  data,
  tlmixture_result,
  plots = TRUE,
  span = 0.9,
  vars_desc = NULL,
  name = "exposures",
  weight_var = NULL,
  round_digits = 4,
  reg_vars = NULL
)
}
\arguments{
\item{data}{tbd}

\item{tlmixture_result}{tbd}

\item{plots}{tbd}

\item{span}{tbd}

\item{vars_desc}{tbd}

\item{name}{tbd}

\item{weight_var}{tbd}

\item{round_digits}{tbd}

\item{reg_vars}{tbd}
}
\description{
Analyze mixture
}
