% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-test-results.R
\name{combine_test_results}
\alias{combine_test_results}
\title{Combine test results}
\usage{
combine_test_results(result, family, verbose = FALSE)
}
\arguments{
\item{result}{tbd}

\item{family}{tbd}

\item{verbose}{tbd}
}
\description{
Combine test results
}
