% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-exposure-weights.R
\name{create_exposure_weights}
\alias{create_exposure_weights}
\title{Create exposure weights}
\usage{
create_exposure_weights(data, outcome, exposures, quantiles,
  verbose = FALSE)
}
\description{
Create exposure weights
}
