% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-mixture.R
\name{display_mixture}
\alias{display_mixture}
\title{Display mixture}
\usage{
display_mixture(
  analysis,
  digits = 3,
  caption = "Mixture analysis",
  label = NULL,
  booktabs = TRUE
)
}
\arguments{
\item{analysis}{tbd}

\item{digits}{tbd}
}
\description{
Display mixture
}
