% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-mixture.R
\name{display_mixture}
\alias{display_mixture}
\title{Display mixture}
\usage{
display_mixture(
  analysis,
  digits = 3,
  caption = "Mixture analysis",
  label = NULL,
  booktabs = TRUE
)
}
\arguments{
\item{analysis}{TLmixture analysis object.}

\item{digits}{Significant digits to round values to; default 3.}

\item{caption}{Caption for the latex table}

\item{label}{Label for the latex output}

\item{booktabs}{Whether to use booktabs output, default True.}
}
\description{
Display mixture
}
