% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-backfit-lm.R, R/mixture-backfit-sl.R
\name{mixture_backfit_lm}
\alias{mixture_backfit_lm}
\title{Create exposure weights via lm with backfitting}
\usage{
mixture_backfit_lm(
  data,
  outcome,
  exposures,
  exposure_groups = NULL,
  max_iterations = 5L,
  quantiles = NULL,
  family = gaussian(),
  verbose = FALSE,
  ...
)

mixture_backfit_lm(
  data,
  outcome,
  exposures,
  exposure_groups = NULL,
  max_iterations = 5L,
  quantiles = NULL,
  family = gaussian(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{tbd}

\item{outcome}{outcome column name}

\item{exposures}{tbd}

\item{exposure_groups}{List of all exposure groups}

\item{quantiles}{tbd}

\item{family}{Not used for this mixture estimator.}

\item{verbose}{tbd}

\item{...}{tbd}
}
\description{
Create exposure weights via lm with backfitting

Create exposure weights via lm with backfitting
}
