% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-backfit-sl3.R
\name{mixture_backfit_sl3}
\alias{mixture_backfit_sl3}
\title{Create exposure weights via SL with backfitting}
\usage{
mixture_backfit_sl3(
  data,
  outcome,
  exposures,
  estimator_mixture = sl3::Lrnr_sl$new(learners = sl3::make_learner(sl3::Stack,
    sl3::make_learner(sl3::Lrnr_mean), sl3::make_learner(sl3::Lrnr_glm)), metalearner =
    sl3::make_learner(sl3::Lrnr_nnls)),
  estimator_confounders = NULL,
  folds_mixture = 5L,
  folds_confounders = 5L,
  exposure_groups = NULL,
  max_iterations = 10L,
  tolerance = 1e-05,
  quantiles = NULL,
  family = "continuous",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{tbd}

\item{outcome}{outcome column name}

\item{exposures}{tbd}

\item{estimator_mixture}{sl3::Lrnr_sl with offset-supporting learners}

\item{estimator_confounders}{sl3::Lrnr_sl with offset-supporting learners}

\item{exposure_groups}{List of all exposure groups}

\item{max_iterations}{tbd}

\item{tolerance}{tbd}

\item{quantiles}{tbd}

\item{family}{Not used for this mixture estimator.}

\item{verbose}{tbd}

\item{...}{tbd}
}
\description{
Create exposure weights via SL with backfitting
}
