% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-pls.R
\name{mixture_pls}
\alias{mixture_pls}
\title{Create exposure weights using partial least squares}
\usage{
mixture_pls(
  data,
  outcome,
  exposures,
  exposure_groups,
  quantiles,
  verbose = FALSE
)
}
\arguments{
\item{data}{tbd}

\item{outcome}{outcome column name}

\item{exposures}{tbd}

\item{exposure_groups}{tbd}

\item{quantiles}{tbd}

\item{verbose}{tbd}
}
\description{
Create exposure weights using partial least squares
}
