% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-backfit-sl3.R
\name{predict.mixture_backfit_sl3}
\alias{predict.mixture_backfit_sl3}
\title{tbd}
\usage{
\method{predict}{mixture_backfit_sl3}(object, data, type = "mixture", ...)
}
\arguments{
\item{object}{tbd}

\item{data}{tbd}

\item{type}{"mixture" (default) or "confounders"}

\item{...}{tbd}
}
\description{
tbd
}
