% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-pls.R
\name{predict.mixture_pls}
\alias{predict.mixture_pls}
\title{prediction for mixture_pls object}
\usage{
\method{predict}{mixture_pls}(obj, data)
}
\arguments{
\item{obj}{tbd}

\item{data}{tbd}
}
\description{
prediction for mixture_pls object
}
